Added logic to fetch from ExerciseDB API and display list of Exercises
Using async/await, data is pulled from the ExerciseDB API and displayed with the ExerciseList component. The ExerciseList component knows which Exercises to display from the props that it receives from its parent component, Exercises. Exercises.jsx utilzes useEffect to obtain the API data and then controls with the power of state. These exercises can be filtered and searched through through input and select elements which are also controlled by state to render the UI.